Add panning functionality
Leaflet has two options: panTo() and setView(). They both do the same thing, but setView has the option for passing zoom levels, if needed. This commit adds that functionality and optional zoom parameters. It no zoom is provided, panTo will default to the the initial zoom level passed in setup.